+++ src/js/defaults.js
@@ -3,7 +3,7 @@
  * @since 2.0.0
  */
 
-//plugin defaults
+// plugin defaults
 const pluginDefaultOptions = {
     // Single snapshot image.
     image: false,
@@ -87,7 +87,10 @@ const pluginDefaultOptions = {
     // Options for animated GIFs using the gifshot library.
     animationOptions: {},
     // Accepts numbers in milliseconds; use this to force intervals-based blobs.
-    timeSlice: 0
+    timeSlice: 0,
+    // Media converter library to use. Legal values are 'ts-ebml' or an empty
+    // string '' to disable (default).
+    convertEngine: ''
 };
 
 export default pluginDefaultOptions;